+++ examples/apollo-enterprise/src/components/dashboard/TaskScheduler.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 // Recharts
 import { ResponsiveContainer } from 'recharts';
 // Lattice
@@ -13,7 +13,7 @@ const data = {
     { title: 'Task D' },
     { title: 'Task E' },
     { title: 'Task F' },
-    { title: 'Task G' }
+    { title: 'Task G' },
   ],
   edges: [
     { source: 'Task A', target: 'Task B' },
@@ -22,20 +22,18 @@ const data = {
     { source: 'Task B', target: 'Task E' },
     { source: 'Task C', target: 'Task D' },
     { source: 'Task E', target: 'Task F' },
-    { source: 'Task E', target: 'Task G' }
-  ]
+    { source: 'Task E', target: 'Task G' },
+  ],
 };
 
-class TaskScheduler extends Component {
-  render() {
-    return (
-      <Widget title="Task Scheduler">
-        <ResponsiveContainer aspect={2}>
-          <Dag {...data} />
-        </ResponsiveContainer>
-      </Widget>
-    );
-  }
-}
+const TaskScheduler = () => {
+  return (
+    <Widget title="Task Scheduler">
+      <ResponsiveContainer aspect={2}>
+        <Dag {...data} />
+      </ResponsiveContainer>
+    </Widget>
+  );
+};
 
 export default TaskScheduler;