+++ examples/apollo-enterprise/src/utils/index.js
@@ -1,5 +1,5 @@
-export const offsetToCursor = offset => Buffer.from(offset.toString()).toString('base64');
+export const offsetToCursor = (offset) => Buffer.from(offset.toString()).toString('base64');
 export const mergeEdges = (left, right) => {
-  const deduplicateEdges = right.filter(edge => left.find(prevEdge => prevEdge.node.id !== edge.node.id));
+  const deduplicateEdges = right.filter((edge) => left.find((prevEdge) => prevEdge.node.id !== edge.node.id));
   return [...left, ...deduplicateEdges];
 };