+++ examples/apollo-real-time-chart/src/Dashboard.js
@@ -17,24 +17,24 @@ import CPUInfo from './components/CPUInfo';
 import DisksUsage from './components/DisksUsage';
 
 // Custom Style
-const styles = theme => ({
+const styles = (theme) => ({
   widget: {
     marginTop: theme.spacing(2),
     marginBottom: theme.spacing(2),
-    padding: theme.spacing(4)
+    padding: theme.spacing(4),
   },
   link: {
-    color: theme.palette.text.secondary
+    color: theme.palette.text.secondary,
   },
   title: {
-    padding: 15
+    padding: 15,
   },
   height400: {
-    height: 400
+    height: 400,
   },
   height300: {
-    height: 300
-  }
+    height: 300,
+  },
 });
 
 class Dashboard extends Component {
@@ -87,8 +87,8 @@ class Dashboard extends Component {
 export default compose(
   graphql(getPlatformInfo, {
     props: ({ data: { getPlatformInfo = {} } }) => ({
-      platformInfo: getPlatformInfo
-    })
+      platformInfo: getPlatformInfo,
+    }),
   }),
   withStyles(styles)
 )(Dashboard);