+++ examples/lattice-ssr/webpack.config.js
@@ -6,40 +6,40 @@ module.exports = [
     module: {
       rules: [
         { test: /\.js$/, exclude: /node_modules/, loader: 'babel-loader' },
-        { test: /\.(ttf|eot|otf|svg|png)$/, loader: 'file-loader' }
-      ]
+        { test: /\.(ttf|eot|otf|svg|png)$/, loader: 'file-loader' },
+      ],
     },
     resolve: {
       alias: {
-        '@material-ui/core/es': '@material-ui/core/es'
-      }
+        '@material-ui/core/es': '@material-ui/core/es',
+      },
     },
     entry: './src/client',
     output: {
-      path: `${__dirname}/public`
+      path: `${__dirname}/public`,
     },
     performance: {
-      hints: false
-    }
+      hints: false,
+    },
   },
   {
     devtool: 'cheap-module-source-map',
     module: {
       rules: [
         { test: /\.js$/, exclude: /node_modules/, loader: 'babel-loader' },
-        { test: /\.(ttf|eot|otf|svg|png)$/, loader: 'file-loader' }
-      ]
+        { test: /\.(ttf|eot|otf|svg|png)$/, loader: 'file-loader' },
+      ],
     },
     resolve: {
       alias: {
-        '@material-ui/core/es': '@material-ui/core/es'
-      }
+        '@material-ui/core/es': '@material-ui/core/es',
+      },
     },
     entry: './src/server',
     target: 'node',
     externals: [nodeExternals()],
     performance: {
-      hints: false
-    }
-  }
+      hints: false,
+    },
+  },
 ];